Took another look at setting env variables.
Slowly refactoring the proto messages to allow each machine to have
their own variables set to faciliate the experiment program running on
diverse systems. Previous setup assumed that all variables were
consistent across all machines which may not be the case. From there, I
verified that the process builder does in-fact set the environment
variables properly.

When a Searcher is initiated, it also receives a copy of the search
experiment information so it can pull out and set all the environmental
variables as appropriate. I believe that after this is all set, the
basic plumbing of deploying the searchers will be established, and I'll
then work on a code commenting session before moving on to storage of
results.