Still messy, but results from the process arrived at the Dispatcher.
Did a good amount of the rewriting that was hinted at yesterday in
Dispatcher and Searcher, and also got the jar entry point fixed from the
earlier stub. The processes do not get cleaned up like they should, but
at least all the info appears to be flowing where it needs to in the one
Searcher case. Next up:
1. Get processes to clean up when complete.
2. Verify multiple replications on a single Searcher.
3. Verify multiple Searchers work.
4. Verify multiple distinct machines work.
Had an idea today: generate experiments according to active learning.